@@ -1,11 +1,20 @@
 apply plugin: 'com.android.application'
+apply plugin: 'kotlin-android'
 
 android {
-    compileSdkVersion 26
+    signingConfigs {
+        config {
+            keyAlias 'androidreleasekey'
+            keyPassword 'third.o5'
+            storeFile file('/Users/saleem/Downloads/thnki.com.jks')
+            storePassword 'third.o5'
+        }
+    }
+    compileSdkVersion 27
     defaultConfig {
         applicationId "com.thnki.queuebreaker"
-        minSdkVersion 19
-        targetSdkVersion 26
+        minSdkVersion 21
+        targetSdkVersion 27
         versionCode 1
         versionName "Auqa"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
