@@ -1,25 +1,36 @@
 package com.thnki.queuebreaker.home.explore;
 
 
-import java.util.List;
+import com.thnki.queuebreaker.home.CollectionListener;
+import com.thnki.queuebreaker.home.CollectionRepository;
 
-class ExplorePresenter implements DataUpdateListener {
+import java.util.ArrayList;
+
+class ExplorePresenter implements CollectionListener<Restaurants> {
+    private static final String TAG = "ExplorePresenter";
     private ExploreView exploreView;
-    private RestaurantsRepository repository;
+    private CollectionRepository<Restaurants> repository;
 
     ExplorePresenter(ExploreView exploreView) {
-
         this.exploreView = exploreView;
-        this.repository = new FirebaseRestaurantsRepository();
-        this.repository.setDataListener(this);
+        repository = new FirestoreRestaurantRepository();
+        repository.addCollectionListener(this, TAG);
     }
 
     @Override
-    public void onDataChanged(List<Restaurants> restaurantsList) {
+    public void onCollectionChanged(ArrayList<Restaurants> restaurantsList) {
         if (restaurantsList.size() > 0) {
-            exploreView.displayRestaurants(repository.getAdapter());
+            exploreView.displayRestaurants(restaurantsList);
         } else {
             exploreView.displayNoRestaurantsFound();
         }
+    }
+
+    public void launchMenu(Restaurants model) {
+        exploreView.launchMenuActivity(model);
+    }
+
+    public void clearListeners() {
+        repository.clearListeners();
     }
 }
