@@ -1,9 +1,10 @@
 package com.thnki.queuebreaker.home.explore;
 
 
-import android.support.v7.widget.RecyclerView;
+import java.util.List;
 
 public interface ExploreView {
-    void displayRestaurants(RecyclerView.Adapter restaurantsList);
+    void displayRestaurants(List<Restaurants> restaurantsList);
     void displayNoRestaurantsFound();
+    void launchMenuActivity(Restaurants model);
 }
